devnet/open_source_software_1_2_Nornir/requirements.txt
hvac==0.10.6
nornir==3.0.0
nornir-jinja2==0.1.1
nornir-netmiko==0.1.1
nornir-utils==0.1.1
cffi>=1.12
cryptography